Remove ugly trick in demo

+++ components/text/src/text.rs
@@ -103,9 +103,10 @@ impl Text {
 
     /// Sets the font size.
     #[must_use]
-    pub fn size(mut self, size: impl IntoSignal<f32>) -> Self {
+    pub fn size(mut self, size: impl IntoSignal<f64>) -> Self {
         // A litle sad we have to do this conversion here
-        let size = size.into_signal().map(|s| s);
+        #[allow(clippy::cast_possible_truncation)]
+        let size = size.into_signal().map(|s| s as f32);
         self.0.content = self
             .0
             .content
